Setup toolbar widget lazy bootstrap
