+++ packages/yii-dev-toolbar/tsconfig.json
@@ -1,7 +1,8 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "rootDir": "./src"
+    "rootDir": "./src",
+    "sourceMap": true
   },
   "include": ["src"],
   "references": [
